Modified precompules & all dependencies. Moved some parametres to engine-types
