+++ engine-precompiles/Cargo.toml
@@ -13,7 +13,6 @@ publish = false
 autobenches = false
 
 [dependencies]
-aurora-engine = { path = "../engine", default-features = false }
 aurora-engine-types = { path = "../engine-types", default-features = false }
 aurora-engine-sdk = { path = "../engine-sdk", default-features = false }
 base64 = { version = "0.13.0", default-features = false, features = ["alloc"] }
@@ -42,5 +41,5 @@ serde_json = "1"
 rand = "0.7.3"
 
 [features]
-contract = ["aurora-engine/contract"]
-log = ["aurora-engine/log"]
+contract = []
+log = []
