+++ engine-precompiles/src/native.rs
@@ -4,8 +4,8 @@ use crate::*;
 use evm::{Context, ExitError};
 #[cfg(feature = "contract")]
 use {
-    aurora_engine::parameters::{PromiseCreateArgs, WithdrawCallArgs},
-    aurora_engine::storage::{bytes_to_key, KeyPrefix},
+    crate::{bytes_to_key, KeyPrefix},
+    crate::{PromiseCreateArgs, WithdrawCallArgs},
     borsh::BorshSerialize,
     evm::backend::Log,
 };
@@ -35,7 +35,7 @@ impl ExitToNear {
     ///
     /// Address: `0xe9217bc70b7ed1f598ddd3199e80b093fa71124f`
     /// This address is computed as: `&keccak("exitToNear")[12..]`
-    pub(crate) const ADDRESS: Address =
+    pub const ADDRESS: Address =
         super::make_address(0xe9217bc7, 0x0b7ed1f598ddd3199e80b093fa71124f);
 }
 
@@ -188,7 +188,7 @@ impl ExitToEthereum {
     ///
     /// Address: `0xb0bd02f6a392af548bdf1cfaee5dfa0eefcc8eab`
     /// This address is computed as: `&keccak("exitToEthereum")[12..]`
-    pub(crate) const ADDRESS: Address =
+    pub const ADDRESS: Address =
         super::make_address(0xb0bd02f6, 0xa392af548bdf1cfaee5dfa0eefcc8eab);
 }
