+++ engine/Cargo.toml
@@ -47,6 +47,7 @@ opt-level = 3
 [dependencies]
 aurora-engine-types = { path = "../engine-types", default-features = false }
 aurora-engine-sdk = { path = "../engine-sdk", default-features = false }
+aurora-engine-precompiles = { path = "../engine-precompiles", default-features = false }
 base64 = { version = "0.13.0", default-features = false, features = ["alloc"] }
 blake2 = { git = "https://github.com/near/near-blake2.git", version = "0.9.1", default-features = false }
 borsh = { version = "0.8.2", default-features = false }
@@ -86,9 +87,9 @@ rand = "0.7.3"
 [features]
 default = ["sha2", "std"]
 std = ["borsh/std", "evm/std", "primitive-types/std", "rlp/std", "sha3/std", "ethabi/std", "logos/std", "bn/std", "aurora-engine-types/std"]
-contract = ["aurora-engine-sdk/contract"]
+contract = ["aurora-engine-sdk/contract", "aurora-engine-precompiles/contract"]
 evm_bully = []
-log = ["aurora-engine-sdk/log"]
+log = ["aurora-engine-sdk/log", "aurora-engine-precompiles/log"]
 meta-call = []
 integration-test = ["log"]
 mainnet = ["contract", "log"]
