+++ engine/src/meta_parsing.rs
@@ -1,3 +1,4 @@
+use crate::precompiles::secp256k1::ecrecover;
 use crate::prelude::*;
 use ethabi::{encode, Token as ABIToken};
 use logos::Logos;
@@ -591,7 +592,7 @@ pub fn parse_meta_call(
     let mut signature: [u8; 65] = [0; 65];
     signature[64] = meta_tx.v;
     signature[..64].copy_from_slice(&meta_tx.signature);
-    match crate::precompiles::ecrecover(H256::from_slice(&msg), &signature) {
+    match ecrecover(H256::from_slice(&msg), &signature) {
         Ok(sender) => {
             result.sender = sender;
             result.input = input;
