+++ engine/src/prelude.rs
@@ -1,6 +1,9 @@
 mod v0 {
+    pub use aurora_engine_precompiles as precompiles;
     pub use aurora_engine_sdk as sdk;
     pub use aurora_engine_sdk::types::*;
+    pub use aurora_engine_types::parameters::*;
+    pub use aurora_engine_types::storage::*;
     pub use aurora_engine_types::types::*;
     pub use aurora_engine_types::*;
     pub use borsh::{BorshDeserialize, BorshSerialize};
@@ -14,7 +17,5 @@ mod v0 {
     pub use crate::log_entry::*;
     pub use crate::parameters::*;
     pub use crate::proof::*;
-    pub use crate::storage;
-    pub use crate::storage::*;
 }
 pub use v0::*;
