+++ etc/state-migration-test/Cargo.toml
@@ -40,3 +40,4 @@ rpath = false
 borsh = { version = "0.8.2", default-features = false }
 aurora-engine = { path = "../../engine", default-features = false, features = ["sha2"] }
 aurora-engine-sdk = { path = "../../engine-sdk", default-features = false }
+aurora-engine-types = { path = "../../engine-types", default-features = false }
